fix: change floats to doubles

+++ tests/test_matrix.c
@@ -5,7 +5,7 @@
 
 MunitResult matrix_test1(const MunitParameter params[], void *fixture)
 {
-	float	initial[4][4] = {
+	double	initial[4][4] = {
 		{    1,    2,    3,    4 },
 		{  5.5,  6.5,  7.5,  8.5 },
 		{    9,   10,   11,   12 },
@@ -26,13 +26,13 @@ MunitResult matrix_test1(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test2(const MunitParameter params[], void *fixture)
 {
-	float	same[4][4] = {
+	double	same[4][4] = {
 		{    1,    2,    3,    4 },
 		{  5.5,  6.5,  7.5,  8.5 },
 		{    9,   10,   11,   12 },
 		{ 13.5, 14.5, 15.5, 16.5 },
 	};
-	float	different[4][4] = {
+	double	different[4][4] = {
 		{    1,    2,    3,    4 },
 		{  5.5,  6.5,  7.5,  8.5 },
 		{    9,   10,   11,   12 },
@@ -53,12 +53,12 @@ MunitResult matrix_test2(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test3(const MunitParameter params[], void *fixture)
 {
-	float	same[4][4] = {
+	double	same[4][4] = {
 		{ 1, 2, 3 },
 		{ 4, 5, 6 },
 		{ 7, 8, 9 },
 	};
-	float	different[4][4] = {
+	double	different[4][4] = {
 		{ 1, 2, 3 },
 		{ 4, 5, 6 },
 		{ 7, 8, 8 },
@@ -78,11 +78,11 @@ MunitResult matrix_test3(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test4(const MunitParameter params[], void *fixture)
 {
-	float	same[4][4] = {
+	double	same[4][4] = {
 		{ 1, 2 },
 		{ 4, 5 },
 	};
-	float	different[4][4] = {
+	double	different[4][4] = {
 		{ 1, 2 },
 		{ 4, 6 },
 	};
@@ -102,19 +102,19 @@ MunitResult matrix_test4(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test5(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]) {
+	t_matrix *a = new_matrix(4, (double[4][4]) {
 			{ 1 , 2 , 3 , 4 },
 			{ 5 , 6 , 7 , 8 },
 			{ 9 , 8 , 7 , 6 },
 			{ 5 , 4 , 3 , 2 },
 			});
-	t_matrix *b = new_matrix(4, (float[4][4]) {
+	t_matrix *b = new_matrix(4, (double[4][4]) {
 			{ -2, 1, 2, 3 },
 			{  3, 2, 1, -1},
 			{  4, 3, 6, 5 },
 			{  1, 2, 7, 8 },
 			});
-	t_matrix *expected = new_matrix(4, (float[4][4]) {
+	t_matrix *expected = new_matrix(4, (double[4][4]) {
 			{ 20, 22 , 50 , 48 },
 			{ 44, 54 , 114, 108},
 			{ 40, 58 , 110, 102},
@@ -132,7 +132,7 @@ MunitResult matrix_test5(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test6(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]) {
+	t_matrix *a = new_matrix(4, (double[4][4]) {
 			{ 1 , 2 , 3 , 4 },
 			{ 2 , 4 , 4 , 2 },
 			{ 8 , 6 , 4 , 1 },
@@ -152,7 +152,7 @@ MunitResult matrix_test6(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test7(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]) {
+	t_matrix *a = new_matrix(4, (double[4][4]) {
 			{ 0 , 1 , 2 , 3 },
 			{ 1 , 2 , 4 , 8 },
 			{ 2 , 4 , 8 , 16 },
@@ -183,14 +183,14 @@ MunitResult matrix_test8(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test9(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float [4][4]) {
+	t_matrix *a = new_matrix(4, (double [4][4]) {
 		{ 0 , 9 , 3 , 0 },
 		{ 9 , 8 , 0 , 8 },
 		{ 1 , 8 , 5 , 3 },
 		{ 0 , 0 , 5 , 8 },
 	});
 
-	t_matrix *expected = new_matrix(4, (float [4][4]) {
+	t_matrix *expected = new_matrix(4, (double [4][4]) {
 		{ 0 , 9 , 1 , 0 },
 		{ 9 , 8 , 8 , 0 },
 		{ 3 , 0 , 5 , 5 },
@@ -218,7 +218,7 @@ MunitResult matrix_test10(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test11(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(2, (float [4][4]){
+	t_matrix *a = new_matrix(2, (double [4][4]){
 			{ 1, 5},
 			{-3, 2},
 		});
@@ -230,12 +230,12 @@ MunitResult matrix_test11(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test12(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(3, (float[4][4]) {
+	t_matrix *a = new_matrix(3, (double[4][4]) {
 			{ 1 , 5 , 0 },
 			{ -3, 2 , 7 },
 			{ 0 , 6 , -3},
 			});
-	t_matrix *expected = new_matrix(2, (float[4][4]) {
+	t_matrix *expected = new_matrix(2, (double[4][4]) {
 			{ -3, 2 },
 			{ 0 , 6 },
 			});
@@ -250,13 +250,13 @@ MunitResult matrix_test12(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test13(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]) {
+	t_matrix *a = new_matrix(4, (double[4][4]) {
 			{ -6 , 1 , 1 , 6 },
 			{ -8 , 5 , 8 , 6 },
 			{ -1 , 0 , 8 , 2 },
 			{ -7 , 1 , -1, 1 },
 			});
-	t_matrix *expected = new_matrix(3, (float[4][4]) {
+	t_matrix *expected = new_matrix(3, (double[4][4]) {
 			{ -6 , 1 , 6 },
 			{ -8 , 8 , 6 },
 			{ -7 ,-1 , 1 },
@@ -272,7 +272,7 @@ MunitResult matrix_test13(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test14(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(3, (float[4][4]) {
+	t_matrix *a = new_matrix(3, (double[4][4]) {
 			{ 3 , 5 , 0 },
 			{ 2 ,-1 ,-7 },
 			{ 6 ,-1 , 5 },
@@ -286,7 +286,7 @@ MunitResult matrix_test14(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test15(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(3, (float[4][4]){
+	t_matrix *a = new_matrix(3, (double[4][4]){
 		{ 3 ,  5 ,  0 },
 		{ 2 , -1 , -7 },
 		{ 6 , -1 ,  5 },
@@ -302,7 +302,7 @@ MunitResult matrix_test15(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test16(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(3, (float[4][4]) {
+	t_matrix *a = new_matrix(3, (double[4][4]) {
 			{ 1 , 2 , 6 },
 			{ -5, 8 , -4},
 			{ 2 , 6 , 4 },
@@ -317,7 +317,7 @@ MunitResult matrix_test16(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test17(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]){
+	t_matrix *a = new_matrix(4, (double[4][4]){
 			{ -2, -8,  3,  5 },
 			{ -3,  1,  7,  3 },
 			{  1,  2, -9,  6 },
@@ -335,7 +335,7 @@ MunitResult matrix_test17(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test18(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]){
+	t_matrix *a = new_matrix(4, (double[4][4]){
 			{ 6 , 4 , 4 , 4 },
 			{ 5 , 5 , 7 , 6 },
 			{ 4 , -9, 3 , -7},
@@ -349,7 +349,7 @@ MunitResult matrix_test18(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test19(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]){
+	t_matrix *a = new_matrix(4, (double[4][4]){
 			{ -4, 2 ,-2 , -3},
 			{ 9 , 6 , 2 , 6 },
 			{ 0 ,-5 , 1 , -5},
@@ -364,14 +364,14 @@ MunitResult matrix_test19(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test20(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]){
+	t_matrix *a = new_matrix(4, (double[4][4]){
 			{-5 ,  2 ,  6 , -8 },
 			{ 1 , -5 ,  1 ,  8 },
 			{ 7 ,  7 , -6 , -7 },
 			{ 1 , -3 ,  7 ,  4 },
 			});
 
-	t_matrix *expected = new_matrix(4, (float[4][4]){
+	t_matrix *expected = new_matrix(4, (double[4][4]){
 			{  0.21805,  0.45113,  0.24060, -0.04511 },
 			{ -0.80827, -1.45677, -0.44361,  0.52068 },
 			{ -0.07895, -0.22368, -0.05263,  0.19737 },
@@ -392,13 +392,13 @@ MunitResult matrix_test20(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test21(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]){
+	t_matrix *a = new_matrix(4, (double[4][4]){
 			{  8, -5,  9,  2 },
 			{  7,  5,  6,  1 },
 			{ -6,  0,  9,  6 },
 			{ -3,  0, -9, -4 },
 			});
-	t_matrix *expected = new_matrix(4, (float[4][4]){
+	t_matrix *expected = new_matrix(4, (double[4][4]){
 			{ -0.15385, -0.15385, -0.28205, -0.53846 },
 			{ -0.07692,  0.12308,  0.02564,  0.03077 },
 			{  0.35897,  0.35897,  0.43590,  0.92308 },
@@ -414,13 +414,13 @@ MunitResult matrix_test21(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test22(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]){
+	t_matrix *a = new_matrix(4, (double[4][4]){
 			{  9 ,  3 ,  0 ,  9 },
 			{ -5 , -2 , -6 , -3 },
 			{ -4 ,  9 ,  6 ,  4 },
 			{ -7 ,  6 ,  6 ,  2 },
 			});
-	t_matrix *expected = new_matrix(4, (float[4][4]){
+	t_matrix *expected = new_matrix(4, (double[4][4]){
 			{ -0.04074 , -0.07778 ,  0.14444 , -0.22222 },
 			{ -0.07778 ,  0.03333 ,  0.36667 , -0.33333 },
 			{ -0.02901 , -0.14630 , -0.10926 ,  0.12963 },
@@ -436,13 +436,13 @@ MunitResult matrix_test22(const MunitParameter params[], void *fixture)
 
 MunitResult matrix_test23(const MunitParameter params[], void *fixture)
 {
-	t_matrix *a = new_matrix(4, (float[4][4]){
+	t_matrix *a = new_matrix(4, (double[4][4]){
 			{  3 , -9 ,  7 ,  3 },
 			{  3 , -8 ,  2 , -9 },
 			{ -4 ,  4 ,  4 ,  1 },
 			{ -6 ,  5 , -1 ,  1 },
 			});
-	t_matrix *b = new_matrix(4, (float[4][4]){
+	t_matrix *b = new_matrix(4, (double[4][4]){
 			{ 8 ,  2 , 2 , 2 },
 			{ 3 , -1 , 7 , 0 },
 			{ 7 ,  0 , 5 , 4 },
